Unstage added changes before aborting
We do git add on the module, so we should clean up on abort. We don't
want to revert anything without asking the user, so just remove from
index.

Prevents accidentally committing rollback changes in a subsequent
changelog commit: if adding changelogs for several modules, you may
hit one that's behind origin and then submit it in the next module's
commit.